13 todo subprocess abstraction

+++ src/vm/vm.ts
@@ -1,90 +1,154 @@
 import { destroyWithLogging } from "@/lib/destroy-with-logging";
-import {
-  createFirecrackerClient,
-  type FirecrackerClient,
-} from "@/lib/firecracker-api";
+import { BindMount, OverlayMount } from "@/lib/file-system";
+import { changePerms } from "@/lib/file-system/utils";
+import { createFirecrackerClient } from "@/lib/firecracker-api";
 import { CrackedError } from "@/lib/judge-error";
-import { baseLogger, registerAsyncProc } from "@/lib/logger";
+import { baseLogger } from "@/lib/logger";
+import { createAsyncProc } from "@/lib/proc/async-proc";
 import { join } from "path";
-import type { VmConfig } from ".";
-import { VmFilesystem } from "./filesys";
-import { VmSocketListener } from "./socket";
 
-const getLogger = (vmId: string) => {
-  const vmLogger = baseLogger.child({}, { msgPrefix: `[${vmId}] ` });
-  return vmLogger;
-};
-
+const logger = baseLogger.child({}, { msgPrefix: "[VM] " });
 const UID = "60000" as const;
 const GID = "60000" as const;
 
-export class VM implements AsyncDisposable {
+class VM implements AsyncDisposable {
   //private stack: AsyncDisposableStack;
 
-  private constructor(
+  private stack: AsyncDisposableStack = new AsyncDisposableStack();
+  private readonly vmProcCmd: string[];
+
+  private readonly depsDir: string;
+  private readonly runDir: string;
+  private readonly jailDir: string;
+
+  private readonly vmRunDir: string;
+  private readonly vmDepsDir: string;
+
+  private readonly guestInitiatedSockPath: string;
+  private readonly hostInitiatedSockPath: string;
+  private readonly firecrackerSockPath: string;
+
+  private isCreated: boolean = false;
+  public isDestroyed: boolean = false;
+
+  constructor(
     public readonly vmId: string,
-    public readonly vmConf: VmConfig,
-    private proc: ReturnType<typeof VM.spawnProcess>,
-    public readonly apiClient: FirecrackerClient,
-    private stack: AsyncDisposableStack,
-  ) {}
+    public readonly vmRoot: string,
+  ) {
+    // host directories
+    this.depsDir = join(vmRoot, "base");
+    this.runDir = join(vmRoot, "run", vmId);
+    // jail directory where VM stuff goes
+    this.jailDir = join(vmRoot, "jail");
+    // directories inside vm chroot
+    this.vmRunDir = join(this.jailDir, "firecracker", vmId, "root", "run");
+    this.vmDepsDir = join(this.jailDir, "firecracker", vmId, "root", "base");
 
-  private static spawnProcess = (vmId: string, vmConf: VmConfig) => {
-    const confFilePath = join("base/", "vm-config.json");
-    const vmLogger = getLogger(vmId);
-    vmLogger.debug(`using config file ${confFilePath}`);
+    this.guestInitiatedSockPath = join(this.runDir, "v.sock_52");
+    this.hostInitiatedSockPath = join(this.runDir, "v.sock");
+    this.firecrackerSockPath = join(this.runDir, "firecracker.socket");
 
-    const proc = Bun.spawn(
-      [
-        vmConf.jailerBinaryPath,
-        "--exec-file",
-        vmConf.firecrackerBinaryPath,
-        "--uid",
-        UID,
-        "--gid",
-        GID,
-        "--id",
-        vmId,
-        "--chroot-base-dir",
-        vmConf.jailDir,
-        "--",
-        "--config-file",
-        // we'll always use an overlayfs to bind to chroot/base for the vm
-        confFilePath,
-      ],
-      {
-        stdout: "pipe",
-        stderr: "pipe",
-      },
-    );
-    // ongoing background process that pipes stdout/stderr to our logger
-    registerAsyncProc({
-      proc,
-      logger: vmLogger,
-    });
-    return proc;
-  };
+    this.vmProcCmd = [
+      //TODO: temp solution
+      join(vmRoot, "jailer"),
+      "--exec-file",
+      join(vmRoot, "firecracker"),
+      "--uid",
+      UID,
+      "--gid",
+      GID,
+      "--id",
+      vmId,
+      "--chroot-base-dir",
+      this.jailDir,
+      "--",
+      "--config-file",
+      join("base", "vm-config.json"),
+    ];
+  }
 
-  static create = async (vmId: string, vmConf: VmConfig): Promise<VM> => {
-    const stack = new AsyncDisposableStack();
+  create = async () => {
+    const vmId = this.vmId;
+    if (this.isCreated) {
+      return;
+    }
+
+    this.isCreated = true;
+
     try {
-      const fs = await VmFilesystem.create(vmId, vmConf);
-      stack.use(fs);
-      const listener = await VmSocketListener.create(fs);
-      stack.use(listener);
+      // we can place sockets in the host dir then bind mount them
+      // into the vm's /run/ directory, as well as any other metadata
+      const run = new BindMount(this.runDir, this.vmRunDir, UID, GID);
+      this.stack.use(run);
+      logger.debug("Created Bind Mount");
 
-      const proc = VM.spawnProcess(vmId, vmConf);
-      const vmLogger = getLogger(vmId).child({ procPid: proc.pid });
-      const api = createFirecrackerClient({
-        socket: fs.firecrackerApiSocketPath,
-        vmId: vmId,
-        fcLogger: vmLogger,
+      // place dependencies in VM chroot dir using overlayfs so it can
+      // modify them ephemerally without cloning the large dependencies in full
+      const deps = new OverlayMount(this.depsDir, this.vmDepsDir);
+      this.stack.use(deps);
+      logger.debug("Created Overlay Mount");
+
+      // change permissions in the overlay; this modifies the overlayfs layer
+      // and allows the VM uid/gid to access the appropriate files we have provided
+      changePerms({
+        path: this.vmDepsDir,
+        uid: UID,
+        gid: GID,
+        mod: "777",
+        recursive: true,
       });
+      logger.debug("Changed perms");
 
-      const vm = new VM(vmId, vmConf, proc, api, stack.move());
-      return vm;
+      const socketProc = await createAsyncProc({
+        cmd: [
+          "socat",
+          `UNIX-LISTEN:${this.guestInitiatedSockPath},fork,reuseaddr`,
+          "TCP:localhost:3000",
+        ],
+        // SIGINT is a graceful exit for `socat` and it will clean up its own
+        // socket file that was created, which isn't the case with SIGTERM
+        killSignal: "SIGINT",
+        logger: logger.child({}, { msgPrefix: `[SOCAT] (vm: ${vmId}) ` }),
+      });
+      this.stack.use(socketProc);
+      logger.debug("Created socket proc");
+
+      const firecrackerSockPath = this.firecrackerSockPath;
+      const vmProc = await createAsyncProc({
+        cmd: this.vmProcCmd,
+        // before destroying the VM, send the ctrl+alt+delete signal
+        // to firecracker, which should cause the VM to gracefully shut down
+
+        async preDestroy(proc) {
+          const api = createFirecrackerClient({
+            socket: firecrackerSockPath,
+            vmId: vmId,
+            fcLogger: logger.child(
+              {},
+              { msgPrefix: `[FIRECRACKER API] (vm: ${vmId}) ` },
+            ),
+          });
+
+          try {
+            await api.PUT("/actions", {
+              body: { action_type: "SendCtrlAltDel" },
+            });
+            // wait for it to be killed for 250ms,
+            // for firecracker that *should* be enough... presumably
+            // if this fails, AsyncProc kills it forcefully
+            await Promise.race([Bun.sleep(250), proc.getExitResult()]);
+          } catch (error) {
+            baseLogger.error(
+              { errorMessage: (error as Error).message },
+              "Firecracker failed in preDestroy!",
+            );
+          }
+        },
+      });
+      this.stack.use(vmProc);
+      logger.debug("Created vm proc");
     } catch (e) {
-      await stack.disposeAsync();
+      await this.destroy();
       throw new CrackedError("VM_CREATE", {
         message: `Failed to create vm ${vmId}`,
         cause: e,
@@ -93,18 +157,15 @@ export class VM implements AsyncDisposable {
   };
 
   destroy = async () => {
+    if (this.isDestroyed) {
+      return;
+    }
+    this.isDestroyed = true;
     await destroyWithLogging(
       async () => {
-        await this.apiClient.PUT("/actions", {
-          body: { action_type: "SendCtrlAltDel" },
-        });
-        this.proc.kill();
         await this.stack.disposeAsync();
-        await this.proc.exited;
       },
-      {
-        label: this.vmId,
-      },
+      { label: this.vmId },
     );
   };
 
@@ -112,3 +173,14 @@ export class VM implements AsyncDisposable {
     await this.destroy();
   }
 }
+
+export const createVm = async (props: { vmId?: string; vmRoot: string }) => {
+  logger.debug("createVm Factory invoked");
+  const { vmId = crypto.randomUUID(), vmRoot } = props;
+  logger.debug({ vmId, vmRoot }, "Creating VM with params:");
+  const vm = new VM(vmId, vmRoot);
+  logger.debug("VM Constructor invoked");
+  await vm.create();
+  logger.debug("VM created");
+  return vm;
+};
